requirements.txt
asgiref==3.7.2
Django==4.2.5
django-appconf==1.0.5
django-ckeditor==6.7.0
django-compressor==4.4
django-cors-headers==4.2.0
django-js-asset==2.1.0
django-libsass==0.8
djangorestframework==3.14.0
libsass==0.22.0
Pillow==10.0.0
pytz==2023.3.post1
rcssmin==1.1.1
rjsmin==1.2.1
sqlparse==0.4.4
typing_extensions==4.7.1
